[HOTSPOT-216] fix: plan - dataAmount int => long으로 변경

+++ src/main/java/hotspot/user/plan/controller/response/PlanResponse.java
@@ -4,7 +4,7 @@ import hotspot.user.plan.domain.DataPeriod;
 
 public record PlanResponse(
         String name,
-        int dataAmount,
+        long dataAmount,
         DataPeriod dataPeriod
 ) {
 }
